requirements.txt
fastapi
uvicorn[standard]
streamlit
python-dotenv
requests
pandas
tqdm
chromadb
sentence-transformers
langchain
huggingface-hub
python-multipart
json